calculate loading from networkStatus
