+++ src/index.ts
@@ -55,7 +55,7 @@ import {
 
 import {
   NetworkStatus,
-} from './queries/store';
+} from './queries/networkStatus';
 
 import {
   ApolloError,
