+++ src/queries/store.ts
@@ -20,25 +20,16 @@ import {
 
 import { isEqual } from '../util/isEqual';
 
+import { NetworkStatus } from './networkStatus';
+
 export interface QueryStore {
   [queryId: string]: QueryStoreValue;
-}
-
-export enum NetworkStatus {
-  loading = 1,
-  setVariables = 2,
-  fetchMore = 3,
-  refetch = 4,
-  poll = 6,
-  ready = 7,
-  error = 8,
 }
 
 export type QueryStoreValue = {
   queryString: string;
   variables: Object;
   previousVariables: Object;
-  loading: boolean;
   networkStatus: NetworkStatus;
   networkError: Error;
   graphQLErrors: GraphQLError[];
@@ -106,7 +97,6 @@ export function queries(
       queryString: action.queryString,
       variables: action.variables,
       previousVariables,
-      loading: true,
       networkError: null,
       graphQLErrors: null,
       networkStatus: newNetworkStatus,
@@ -132,12 +122,11 @@ export function queries(
 
     newState[action.queryId] = {
       ...previousState[action.queryId],
-      loading: false,
       networkError: null,
       graphQLErrors: resultHasGraphQLErrors ? action.result.errors : null,
       previousVariables: null,
       networkStatus: NetworkStatus.ready,
-    } as QueryStoreValue;
+    };
 
     return newState;
   } else if (isQueryErrorAction(action)) {
@@ -154,10 +143,9 @@ export function queries(
 
     newState[action.queryId] = {
       ...previousState[action.queryId],
-      loading: false,
       networkError: action.error,
       networkStatus: NetworkStatus.error,
-    } as QueryStoreValue;
+    };
 
     return newState;
   } else if (isQueryResultClientAction(action)) {
@@ -169,7 +157,6 @@ export function queries(
 
     newState[action.queryId] = {
       ...previousState[action.queryId],
-      loading: !action.complete,
       networkError: null,
       previousVariables: null,
       // XXX I'm not sure what exactly action.complete really means. I assume it means we have the complete result
@@ -205,7 +192,6 @@ function resetQueryState(state: QueryStore, action: StoreResetAction): QueryStor
     // XXX set loading to true so listeners don't trigger unless they want results with partial data
     res[key] = {
       ...state[key],
-      loading: true,
       networkStatus: NetworkStatus.loading,
     };
