+++ test/client.ts
@@ -31,7 +31,7 @@ import {
   print,
 } from 'graphql-tag/printer';
 
-import { NetworkStatus } from '../src/queries/store';
+import { NetworkStatus } from '../src/queries/networkStatus';
 
 import {
   createStore,
@@ -472,7 +472,6 @@ describe('client', () => {
         '1': {
           queryString: print(query),
           variables: {},
-          loading: false,
           networkStatus: NetworkStatus.ready,
           networkError: null,
           graphQLErrors: null,
