+++ test/scheduler.ts
@@ -8,7 +8,7 @@ import {
   createApolloStore,
 } from '../src/store';
 import mockNetworkInterface from './mocks/mockNetworkInterface';
-import { NetworkStatus } from '../src/queries/store';
+import { NetworkStatus } from '../src/queries/networkStatus';
 import gql from 'graphql-tag';
 
 describe('QueryScheduler', () => {
