+++ test/util/subscribeAndCount.ts
@@ -7,11 +7,17 @@ import wrap from './wrap';
 export default function(done: MochaDone, observable: ObservableQuery<any>,
     cb: (handleCount: number, result: ApolloQueryResult<any>) => any): Subscription {
   let handleCount = 0;
-  return observable.subscribe({
-    next: wrap(done, result => {
-      handleCount++;
-      cb(handleCount, result);
-    }),
+  const subscription = observable.subscribe({
+    next: result => {
+      try {
+        handleCount++;
+        cb(handleCount, result);
+      } catch (e) {
+        subscription.unsubscribe();
+        done(e);
+      }
+    },
     error: done,
   });
+  return subscription;
 };
